Updated to Running Version - TODO get RFID working with Tap valves

+++ install/includes/configprocessor.php
@@ -3,6 +3,17 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <title>Installation Processor</title>
+
+<script>
+window.onpageshow = function(evt) {
+    // If persisted then it is in the page cache, force a reload of the page.
+    if (evt.persisted) {
+        document.body.style.display = "none";
+        location.reload();
+    }
+};
+</script>
+
 </head>
 <body>
 <?php
@@ -13,14 +24,18 @@ require_once __DIR__.'/sql_parse.php';
 
 //Process and load form data
 $servername = $_POST["servername"];
+$rootuser = $_POST["rootuser"];
 $rootpass = $_POST["rootpass"];
-$dbuser = $_POST["dbuser"];
-$dbpass1 = $_POST["dbpass1"];
-$dbpass2 = $_POST["dbpass2"];
+$databasename = $_POST["database"];
+$dbuser = $rootuser;//$_POST["dbuser"];
+$dbpass1 = $rootpass;//$_POST["dbpass1"];
+$dbpass2 = $rootpass;//$_POST["dbpass2"];
 $adminuser = $_POST["adminuser"];
 $adminpass1 = $_POST["adminpass1"];
 $adminpass2 = $_POST["adminpass2"];
 $action = $_POST["selectaction"];
+$adminname = $_POST["adminname"];
+$adminemail = $_POST["adminemail"];
 
 //Create the MD5 hash value for the admin password
 $adminhash = md5($adminpass1);
@@ -47,12 +62,12 @@ flush();
 //Validate DB connectivity
 echo "Checking DB connectivity...";
 flush();
-$con=mysqli_connect($servername,"root",$rootpass);
+$mysqli = new mysqli($servername,$rootuser,$rootpass);
 
 if (mysqli_connect_errno())
-  {
-  $validerror .= "<br><strong>Cannot connect the the database using the supplied information.</strong>";
-  }
+{
+$validerror .= "<br><strong>Cannot connect the the database using the supplied information.</strong>";
+}
 echo "Success!<br>";
 flush();
 
@@ -61,17 +76,18 @@ echo "Checking config folder permissions...";
 flush();
 if (!is_writable(dirname('../../includes/functions.php')))
 {
-   $validerror .= "<br><strong>Cannot write the configuration files. Please check the /includes/ folder permissions. See the RPints Installation page on www.raspberrypints.com.</strong>";
+$validerror .= "<br><strong>Cannot write the configuration files. Please check the /includes/ folder permissions. See the RPints Installation page on www.raspberrypints.com.</strong>";
 }
 
 if (!is_writable(dirname('../../admin/includes/checklogin.php')))
 {
 $validerror .= "<br><strong>Cannot write the configuration files. Please check the /admin/includes/ folder permissions. See the RPints Installation page on www.raspberrypints.com.</strong>";
 }
+//$mysqli->close();
 echo "Success!<br>";
 flush();
 
-  //##TODO## Check if administrator account already exists
+//##TODO## Check if administrator account already exists
 
 
 
@@ -84,30 +100,40 @@ if ($validerror !='')
 		echo "</body></html>";
 		die();
 	}
-
+// CLEAR INSTALLATION DATA ROUTINES
 if ($action == 'remove')
 {
-	echo "Deleting raspberrypints database...";
+	echo "Deleting " + $databasename + " database...";
 	flush();
-	$con=mysqli_connect($servername,"root",$rootpass);
+	//$mysqli = new mysqli($servername,$rootuser,$rootpass);
 	// Check connection
 
 	if (mysqli_connect_errno())
-	  {
-	  echo "Failed to connect to MySQL: " . mysqli_connect_error();
-	  }
+	{
+		echo "Failed to connect to MySQL: " . $mysqli->connect_error();
+	}
 
-	$sql = "DROP database raspberrypints;";
-	$result = mysqli_query($con,$sql);
-	mysqli_close($con);
+	$sql = "DROP database " . $databasename . ";";
+	$result = $mysqli->query($sql);
+	//$mysqli->close();
 	echo "Success!<br>";
 	flush();
 	
 	echo "Removing configuration files...";
 	flush();
-	unlink('../../includes/config.php');
-	unlink('../../admin/includes/conn.php');
-	unlink('../../admin/includes/configp.php');
+	try {
+	
+	if (file_exists('../../includes/config.php')) unlink('../../includes/config.php');
+	if (file_exists('../../includes/conn.php'))   unlink('../../admin/includes/conn.php');
+	//if (file_exists('../../includes/configp.php'))unlink('../../admin/includes/configp.php');
+	} catch (Exception $e) {
+		echo 'Caught exception: ',  $e->getMessage(), "\n";
+	}
+	
+	//unlink('../../includes/config.php');
+	//unlink('../../admin/includes/conn.php');
+	//unlink('../../admin/includes/configp.php');
+
 	echo "Success!<br>";
 	flush();
 }
@@ -130,24 +156,39 @@ require_once __DIR__.'/config_files.php';
 	flush();
 
 	file_put_contents('../../admin/includes/conn.php', $adminconfig1);
-	file_put_contents('../../admin/includes/configp.php', $adminconfig2);
+	//file_put_contents('../../admin/includes/configp.php', $adminconfig2);
 	
 	echo "Success!<br>";
 	flush();
+
+	//-----------------Create RPints Database----------------------
+	echo "Creating RPints database...";
+	flush();
+
+	$sql = "DROP DATABASE `" . $databasename . "`;";
+	$result = $mysqli->query($sql);
+	$sql = "CREATE DATABASE `" . $databasename . "` DEFAULT CHARACTER SET latin1 COLLATE latin1_swedish_ci;";
+	$result = $mysqli->query($sql);
+	$sql = "USE `" . $databasename . "`;";
+	$result = $mysqli->query($sql);
+	//mysqli_close($con);
+	echo "Success!<br>";
+	flush();
+
 	//-----------------Create RPints User--------------------------
 	echo "Creating RPints database user...";
 	flush();
-	$con=mysqli_connect($servername,"root",$rootpass);
+	//$mysqli = new mysqli($servername,$rootuser,$rootpass);
 	// Check connection
 
 	if (mysqli_connect_errno())
-	  {
-	  echo "Failed to connect to MySQL: " . mysqli_connect_error();
-	  }
+	{
+		echo "Failed to connect to MySQL: " . $mysqli->connect_error();
+	}
 
 	$sql = "GRANT ALL ON *.* TO '" . $dbuser . "'@'" . $servername . "' IDENTIFIED BY '" . $dbpass1 . "' WITH GRANT OPTION;";
-	$result = mysqli_query($con,$sql);
-	mysqli_close($con);
+	$result = $mysqli->query($sql);
+	//$mysqli->close();
 	echo "Success!<br>";
 	flush();
 
@@ -164,36 +205,52 @@ require_once __DIR__.'/config_files.php';
 	$sql_query = split_sql_file($sql_query, ';');
 
 
-	mysql_connect($servername,'root',$rootpass) or die('error connection');
+	//$mysqli = new mysqli($servername,$rootuser,$rootpass) or die('error connection');
 
 	$i=1;
 	foreach($sql_query as $sql){
-	//echo $i++;
-	//echo "	";
-	//echo $sql;
-	//echo "<br>";
-	mysql_query($sql) or die('error in query');
+		if(rtrim($sql) == "") continue;
+		//echo "	";
+		//echo $sql;
+		//echo "<br>";
+		$mysqli->query($sql) or die('error in query '.$i.'['.substr($sql,0,80).'] ['.$mysqli->error.']');
+		//echo "<br>";
+		$i++;
 	}
 
+	//$mysqli->close();
 	echo "Success!<br>";
 	flush();
 
 	//-----------------Add the admin user to the Users DB----------
 	echo "Adding new admin user...";
 	flush();
-	$con=mysqli_connect($servername,"root",$rootpass,"raspberrypints");
+	////$mysqli = new mysqli($servername,$rootuser,$rootpass,"raspberrypints");
 	// Check connection
 
 	if (mysqli_connect_errno())
-	  {
-	  echo "Failed to connect to MySQL: " . mysqli_connect_error();
-	  }
+	{
+	echo "Failed to connect to MySQL: " . $mysqli->connect_error();
+	}
 	$currentdate = Date('Y-m-d H:i:s');
-	$sql = "INSERT INTO users (username, password, name, email, createdDate, modifiedDate) VALUES ('" . $adminuser . "','" . $adminhash . "','name','email','" . $currentdate . "','" . $currentdate . "');";
-	$result = mysqli_query($con,$sql);
-	mysqli_close($con);
+	$sql = "INSERT INTO users (username, password, name, email, createdDate, modifiedDate) VALUES ('" . $adminuser . "','" . $adminhash . "','" . $adminname . "','" . $adminemail . "','" . $currentdate . "','" . $currentdate . "');";
+	$result = $mysqli->query($sql);
+	//$mysqli->close();
 	echo "Success!<br>";
 	flush();
+	
+	//-----------------Delete the index.html page, if it exists -----------------
+	$index = '../../index.html';
+	echo "Deleting default Apache index...";
+	flush();
+	if (file_exists($index)) {
+		unlink($index);
+		echo "Success! <br>";
+	}	else {
+		echo "Success! File already deleted <br>";
+	}
+	flush();
+	
 	//-----------------Load the sample data if requested-----------
 
 		if(!empty($_POST['sampledata'])) 
@@ -211,28 +268,34 @@ require_once __DIR__.'/config_files.php';
 			$sql_query = split_sql_file($sql_query, ';');
 
 
-			mysql_connect($servername,'root',$rootpass) or die('error connection');
+			////$mysqli = new mysqli($servername,$rootuser,$rootpass) or die('error connection');
 
 			$i=1;
 			foreach($sql_query as $sql){
-			//echo $i++;
-			//echo "	";
-			mysql_query($sql) or die('error in query');
+				if(rtrim($sql) == "") continue;
+				//echo "	";
+				//echo $sql;
+				//echo "<br>";
+				$mysqli->query($sql) or die('error in query '.$i.'['.substr($sql,0,80).'] ['.$mysqli->error.']');
+				//echo "<br>";
+				$i++;
 			}
 
 			
+			//$mysqli->close();
 			echo "Success!<br>";
 			flush();
 		}
 }
+$mysqli->close();
 
 
 if ($action != 'remove')
 {
 	##TODO## Add better error handling before showing the Success message
 	echo '<br /><br /><br /><h3> Congratulations! Your Raspberry Pints has been setup successfully.<br />';
-	echo 'Tap List - <a href="http://' . $_SERVER['HTTP_HOST'] . '/index.php">http://' . $_SERVER['HTTP_HOST'] . '/index.php</a><br />';
-	echo 'Administration - <a href="http://' . $_SERVER['HTTP_HOST'] . '/admin/index.php">http://' . $_SERVER['HTTP_HOST'] . '/admin/index.php</a><br />';
+	echo 'Click for - <a href="../../index.php">Tap List</a><br />';
+	echo 'Click for - <a href="../../admin/index.php">Administration </a><br />';
 }
 
 ?>
